# Conflicts: #	src/main/java/com/example/demo/Config/SecurityConfig.java #	src/main/java/com/example/demo/Config/URLPattern.java #	src/main/java/com/example/demo/Controller/PostController.java #	src/main/java/com/example/demo/DTO/CommentPageDto.java #	src/main/java/com/example/demo/Service/Impl/PostServiceImpl.java #	src/main/java/com/example/demo/Service/Impl/UserServiceImpl.java #	src/main/java/com/example/demo/Service/PostService.java #	src/test/java/com/example/demo/Controller/PostControllerTest.java #	src/test/java/com/example/demo/Repository/CommentLikeRepositoryTest.java #	src/test/java/com/example/demo/Service/PostServiceTest.java #	src/test/java/com/example/demo/Service/UserServiceTest.java

+++ src/main/java/com/example/demo/Repository/CommentLikeRepository.java
@@ -13,5 +13,5 @@ public interface CommentLikeRepository extends JpaRepository<CommentLikeEntity, 
 
     List<CommentLikeEntity> findByCidAndLikes(CommentEntity cid, Boolean likes);
 
-    int countByCidAndLikes(CommentEntity cid, Boolean likes);
+    Long countByCidAndLikes(CommentEntity cid, Boolean likes);
 }
